pyproject.toml
[project]
name = "build-your-own-ai"
version = "0.1.0"
description = "build your own ai."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "cozepy>=0.19.0",
    "dashscope>=1.24.2",
    "fastapi>=0.116.1",
    "flask>=3.1.2",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.6",
    "missingno>=0.5.2",
    "modelscope>=1.29.1",
    "openai>=1.102.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "ydata-profiling>=4.6.0",
    "peft>=0.17.1",
    "plotly>=6.3.1",
    "prophet>=1.1.7",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.4.1,<1.14",
    "seaborn>=0.10.1,<0.13",
    "statsmodels>=0.13.2,<0.14",
    "torch>=2.8.0",
    "transformers>=4.56.0",
    "uvicorn>=0.35.0",
    "xgboost>=3.0.5",
    "mlxtend>=0.23.4",
    "bayesian-optimization>=2.0.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "uv>=0.9.0",
]
